feature: popup context
added popup context so you can trigger popups more easily. a popup is now a portal and is connected to its own div in the index.html. furthermore with the popup context you can trigger a popup with the method showPopup now and it will appear on top no matter where you navigate or on which page you are